% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{as.edge_list.interaction_table}
\alias{as.edge_list.interaction_table}
\alias{as.edgelist.interaction_table}
\title{Create edge list from interaction table}
\usage{
\method{as.edgelist}{interaction_table}(table)
}
\arguments{
\item{table}{An \code{interaction_table}}
}
\value{
The result of this function can be fed directly into \link{igraph} for making interacting network
}
\description{
Converts an interaction table to a two column character matrix
where each row represent an edge between the OTUs. The entities in the each row are the name of the
two interacting OTUs.
}
