% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{as.edgelist}
\alias{as.edgelist}
\title{Convert an object to a list of edges}
\usage{
as.edgelist(x, ...)
}
\arguments{
\item{x}{The R object to convert to an edgelist}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This is a S3 generic to convert a suitable object represetning (on some way) a graph to an
egelist which can be inported to \code{igraph} by the function \link{graph_from_edgelist}.
}
