% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{autoplot.cvLV}
\alias{autoplot.cvLV}
\title{Create cross-validation colorplot}
\usage{
\method{autoplot}{cvLV}(object, target = "RMSE", ...)
}
\arguments{
\item{object}{A \code{cvLV} object returned from \code{\link{cv.LV}}}

\item{target}{The cross-validation target to pick, one of \code{'RMSE'} or \code{'MAE'}}

\item{...}{other arguments passed to methods}
}
\value{
A \link{ggplot} colorplots cross-validation errors for the different weights combination
}
\description{
This function views the cross-validation error in a colorplot as a function
of the regularization weights. Hence, this approach is suitable to detect whether the cross-validation
procedure contains a reasonable optimum.
}
\examples{
library(micInt)
library(phyloseq)
data("seawater")
physeq_list <- subdivide_by_environment(seawater,"Reactor")
time_series <- lapply(physeq_list$phyloseq,OTU_time_series,
time_points ="Week")
cv_res <- cv.LV(time_series,n_folds = 3,
kind = "integral",
weights = expand.grid(self= c(1,2),
interaction = c(1,2)
)
)
autoplot(cv_res, target = "MAE")
}
