% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_pipeline.R
\name{autoplot.interaction_table}
\alias{autoplot.interaction_table}
\title{autoplot.interaction_table}
\usage{
\method{autoplot}{interaction_table}(
  object,
  OTU_stat,
  type = "num_int",
  cutoff_type = "q",
  abundance_type = "mean",
  ...
)
}
\arguments{
\item{object}{An \code{interaction_table}}

\item{OTU_stat}{OTU statistics obtained from the function \link{OTU_stats}}

\item{type}{One of the following: \itemize{
\item \code{'num_int'} Plots the number of interactions for each OTU against
its abundance
\item \code{'ab_prod'} Plots the product of the abundances of each significant pairs of
OTUs against the q- or p-value of the interaction
\item \code{'sim_cor'} Plots the absolute value of \code{sim.score} of each significant interactions against
the q- or p-value of the interaction
}}

\item{cutoff_type}{One of the following:
\itemize{
\item \code{'q'} Plots according to the q-values
\item \code{'p'} Plots according to the p-values
}
This parameter is ignored if \code{type='num_int'}}

\item{abundance_type}{The type of abundance to use in the plots. One of \code{c('mean','median','max')}.}

\item{...}{Currently ignored}
}
\value{
A \link{ggplot} object showing the desired diagnostic plot

#' @examples
library(micInt)
data("seawater")
sim.scores <- similarity_measures(subset= c("spearman","pearson"))
res <- runAnalysis(OTU_table = seawater,sim.scores = sim.scores,returnVariables = 'ccrepe_res',iterations = 100,parallel = FALSE)
int_table <- create_interaction_table(res$ccrepe_res$spearman$res)
stats <- OTU_stats(seawater)
autoplot(int_stable,stats,type = "ab_prod",cutoff_type = "p",abundance_type = "max")
}
\description{
Makes a diagnostic plot of an \code{interaction_table}
}
\details{
Note that if \code{type='sim_cor'}, a column named \code{sign}, indicating the sign of the interaction,
is available even though the plotting function does not use this feature.
}
