% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccrepe_analysis.R
\name{ccrepe_analysis}
\alias{ccrepe_analysis}
\title{Conduct ccrepe analysis}
\usage{
ccrepe_analysis(ccrepe_job, commonargs, parallel = TRUE, verbose = TRUE)
}
\arguments{
\item{ccrepe_job}{A list of jobs to be passed to \code{ccrepe}. The list themselves are named lists with the arguments
being passed to \code{\link{ccrepe}}}

\item{commonargs}{\code{ccrepe} arguments common for all jobs}

\item{parallel}{Should the jobs be run in parallel?}

\item{verbose}{Should the function display how much time it spent?}
}
\value{
A list of the results of the various jobs.
Each element of this list containings the ccrepe results in addition to the outputargs passed as the jobs
}
\description{
A wrapper around the \code{ccrepe} function, provides parallel analysis
}
\seealso{
\link{ccrepe}
}
