% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{collapse_taxonomy}
\alias{collapse_taxonomy}
\title{Collapse taxonomy into single strings}
\usage{
collapse_taxonomy(phyloseq)
}
\arguments{
\item{phyloseq}{A \code{\link{phyloseq}} object containing the \code{tax_table} slot. If this slot does not exist,
the function will return \code{NULL}.}
}
\value{
A named character vector, where the names are the OTU names and the values are the collapsed taxonomies.
}
\description{
Given a phyloseq object with a taxonomy table, this function extracts the taxonomy of each OTU and
collapse the result into a single string
}
