% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTU_stats.R
\name{countInteractions}
\alias{countInteractions}
\title{countInteractions}
\usage{
countInteractions(IDs, interactions_table)
}
\arguments{
\item{IDs}{The name of each OTU}

\item{interactions_table}{An \code{interaction_table}}
}
\description{
Counts the number of interactions for each OTU
}
