% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_measures.R
\name{create_ccrepe_jobs}
\alias{create_ccrepe_jobs}
\title{create_ccrepe_jobs}
\usage{
create_ccrepe_jobs(data = NULL, sim.scores = similarity_measures(),
  ccrepe_defaultargs = NULL, prefix = "significant_interactions",
  postfix = ".csv")
}
\description{
Creates ccrepe jobs out of sim.score functions
}
