% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{cut_abundances}
\alias{cut_abundances}
\title{cut_abundances}
\usage{
cut_abundances(
  refined_table,
  abundance_cutoff = 0,
  cutoff_type = "mean",
  renormalize = TRUE
)
}
\arguments{
\item{refined_table}{A \code{data.frame} where OTUs are in columns and samples in rows with metadata removed}

\item{abundance_cutoff}{Numeric, the threshold cutoff value. If it is \code{NULL}, the filtering process is skipped.}

\item{cutoff_type}{The type of measure to base the cutoff on. Can be any
of \code{'mean'}, \code{median}, \code{max} which cuts away
OTUs based on mean, median and maximum abundance, repectivly}

\item{renormalize}{Logical, should the abundances be renormalized after the procedyre?}
}
\description{
Removes metadata columns from dataset
}
