% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{cv.LV_fit}
\alias{cv.LV_fit}
\title{Cross-validate fit for Lotka-Volterra coefficients}
\usage{
cv.LV_fit(equations, n_folds, weigths = expand.grid(self = 0.1 * 0:100,
  interaction = 0.1 * 0:100))
}
\arguments{
\item{equations}{A list of equation systems returned from \code{\link{integralSystem}}}

\item{n_folds}{The number of folds to apply}

\item{weights}{A n times 2 data.frame containing the weights to be tried out. The two colmns are: \itemize{
\item \code{'self'} The regularization for the maximal growth rate
\item \code{'interaction'} The regulartization for the interactions between OTUs
}
If missing, the cross validation is carried out over a quadratic grid from 0 to 10 in steps of 0.1}
}
\value{
A data.frame containing the parameter values, and the two columns:
 \itemize{
\item \code{'RMSE'} The root mean square error of the right sides
\item \code{'MAE'} The mean absolute error of the right sides
}
}
\description{
The function preforms k-fold cross-validation on a grid of regularization parameters
}
