% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{edgelist}
\alias{edgelist}
\alias{as.edgelist}
\title{Convert an object to a list of edges}
\usage{
as.edgelist(x, ...)
}
\arguments{
\item{x}{The R object to convert to an edgelist}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This is a S3 generic to convert a suitable object represetning (on some way) a graph to an
edgelist which can be inported to \code{igraph} by the function \link{graph_from_edgelist}.
}
\examples{
library(micInt)
library(igraph)
sim.scores <- similarity_measures(subset= c("spearman","pearson"))
res <- runAnalysis(OTU_table = seawater,sim.scores = sim.scores,parallel = FALSE)
int_table <- create_interaction_table(res[[1]]$ccrepe_res)
edgelist <- as.edgelist(int_table)
graph_from_adj_list(edgelist)
}
