% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integralSystem.R
\name{integralSystem}
\alias{integralSystem}
\title{Make Lotka-Volterra coefficient system}
\usage{
integralSystem(time_series, kind = "integral", removeZeros = TRUE)
}
\arguments{
\item{time_series}{An \code{OTU_time_series} or a list of them
If a list is passed, the OTUs must be the same and in the same order in each of the sub-objects.}

\item{kind}{Charachter, one of \code{c('integral','log_integral')} Choose whether use the integral or log-integral approach
described in the article cited.}

\item{removeZeros}{Logical, for the \code{'integral'} option, remove the equations which
produce zero rows. This is ignored for the \code{'log_integral'} because in that
case, all two consecutive measurements which contains a zero for the OTU in question,
must be discared anyway.}
}
\value{
A list of lists (one for each OTU) with two elements:
\itemize{
\item \code{A} The coefficient matrix for the system
\item \code{b} The right side of the system
}
}
\description{
This function accepts time series of OTU measurements
and the linear equation system for its coefficients, as described in cited article.
}
\references{
P. H. Kloppers and J. C. Greeff. ``Lotka-Volterra model parameter
estimation using experiential data''. In: \emph{Appl. Math. Comput. 224}
(Nov. 2013), pp. 817–825. ISSN: 0096-3003. DOI: \url{https://doi.org/10.1016/j.amc.2013.08.093}
}
