% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_measures.R
\name{mean_scale}
\alias{mean_scale}
\title{mean_scale}
\usage{
mean_scale(sim.scores = similarity_measures(), append = TRUE)
}
\arguments{
\item{sim.scores}{A list of \code{sim.measure} objects}

\item{append}{Logical, should the \code{sim.scores} passed to the function also be returned?}
}
\description{
Creates \code{sim.measure} objects which devide each compontent of by its mean before performing
the calculations
#' @examples
library(micInt)
sim_measures <- similarity_measures(subset = c("pearson","spearman","jaccard","bray_curtis))
mean_scaled_sim_measures <- noisify(sim_measures,append = FALSE)
mean_scaled_sim_measures
The Bray-Curtis similarity is the only one returned due to its properties
mean_scaled_sim_measures_append <- noisify(sim_measures,append = TRUE)
mean_scaled_sim_measures_append
}
