% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{micInt-package}
\alias{micInt}
\alias{micInt-package}
\title{micInt: Find microbial interactions}
\description{
This packages utilizes different approaches to find and analyze microbial interactions. 
    The CoNet approach utilzes similarity measures and 
    bootstrapping-permutation methology to find significant correlations between different 
    OTUs in a microbiome data set. The analysis are automized and run in parallel for
    convenient comparison of different similarity measures and methods. The Lotka-Volterra approach is
    based on time series and estimates the Lotka-Volterra coefficients.
}
\details{
\subsection{Common}{
The starting point for this packages is an OTU table. See description of its format under \code{\link{runAnalysis}}. Note the functions
\code{\link{OTU_stats}} and \code{\link{refine_data}} which are used internally by many of the functions, but
can be useful in user-written code as well.
}

\subsection{CoNet}{
For the CoNet approach, the major function is \code{\link{runAnalysis}} which provides a more or less
complete pipeline. This is a wrapper around the package \code{\link{ccrepe}}. For more advanced users, the lower level
API using \code{\link{ccrepe_analysis}} might be preferable. In addition, there are many auxillary functions
such as generating interaction tables (\code{\link{create_interaction_table}}), writing them to file (\code{\link{write.interaction_table}})
and making diagnostic plots of the tables (\code{\link{autoplot.interaction_table}}).
}

\subsection{Lotka-Volterra}{
This approach is based on
The Lotka-Volterra systems are generated by \code{\link{integralSystem}}
For the time being, \code{\link{ridge_fit}} is the only function available to solve
the systems (works even if they are over- or underdetermined) by ridge regularization.
The regularization parameters have to be determined somehow, this is where \code{\link{cv.LV}} comes
into play. It performs cross-validation in order to pick the best regularization parameters.
Finally, the estimated coefficients can be put into the differential equations again in order
to predict future dynamics of a microbial community using \code{\link{predict.LV}}
}
}
\references{
Faust K and Raes J. CoNet app: inference of biological association networks using Cytoscape \emph{F1000Research 2016}, 5:1519
(doi: \url{https://doi.org/10.12688/f1000research.9050.2})

Karoline Faust et al. Microbial Co-occurrence Relationships in the Human Microbiome".
In: \emph{PLoS Comput. Biol.} 8.7 (July 2012). issn: 1553-734X. doi: \url{https://doi.org/10.1371/journal.pcbi.
1002606}.


Richard R. Stein et al. ``Ecological Modeling from Time-Series Inference: Insight into Dy-
namics and Stability of Intestinal Microbiota.'' In: \emph{PLoS Comput. Biol. 9.12 (desember
2013)}. issn: 1553-7358. doi: \url{https://doi.org/10.1371/journal.pcbi.1003388}.

P. H. Kloppers and J. C. Greeff. ``Lotka-Volterra model parameter
estimation using experiential data''. In: \emph{Appl. Math. Comput. 224}
(Nov. 2013), pp. 817–825. ISSN: 0096-3003. DOI: \url{https://doi.org/10.1016/j.amc.2013.08.093}
}
\seealso{
\code{\link{refine_data}} Convert an OTU table into a refined table (required for some functions in the package)

\code{\link{OTU_stats}} Get statistics on OTUs

\code{\link{runAnalysis}} High level CoNet wrapper

\code{\link{ccrepe_analysis}} Low level CoNet wrapper

\code{\link{create_interaction_table}} Make CoNet interaction table

\code{\link{integralSystem}} Make equations for estimating the Lotka-Volterra coeffcients

\code{\link{ridge_fit}} Solve Lotka-Volterra systems by ridge regularization

\code{\link{cv.LV}} Find optimal regularization parameters by cross-validation

\code{\link{predict.LV}} Predict the development of a microbial community by
estimated Lotka-Volterra coeffcients
}
\keyword{internal}
