% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_measures.R
\name{noisify}
\alias{noisify}
\title{Noisify}
\usage{
noisify(sim.scores = similarity_measures(), magnitude = 1e-05,
  noise = c("none", "uniform", "normal"))
}
\arguments{
\item{sim.scores}{The list of similarity measures to nosify}
}
\value{
A list of \code{sim.score} functions corresponding to the input where noise have been
added
}
\description{
Creates similarity measure functions which adds random noise
at certain magnitude to the data before calculating
the similarity score
}
