% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTU_stats.R
\name{OTU_stats}
\alias{OTU_stats}
\title{OTU statistics}
\usage{
OTU_stats(OTU_table, metadataCols = c("OTU Id", "taxonomy"))
}
\arguments{
\item{OTU_table}{The raw OTU table}

\item{metadataCols}{See the corresponding description for \link{runAnalysis}}
}
\value{
A \code{data.frame} with one row for each OTU with the following
elements:

\itemize{
\item \code{ID} The OTU name
\item \code{meanAbundance}
\item \code{medianAbundance}
\item \code{maxAbundance}
\item \code{numberNonZero} The number of samples where the OTU has a
non-zero abundance
\item \code{taxonomy}
}
}
\description{
This function returns a summary of OTU statistics for the OTUs of
an OTU table
}
