% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_class.R
\docType{class}
\name{OTU_time_series}
\alias{OTU_time_series}
\alias{OTU_time_series}
\title{Time series class for OTU tables}
\usage{
OTU_time_series(table, time_points)
}
\description{
This class provides a way to integrate measurements of time
to OTU tables, useful for time series analysis
}
\section{Slots}{

\describe{
\item{\code{table}}{The refined OTU table returned from \link{refine_data}}

\item{\code{time_points}}{Numeric, the time (measured in secounds) from the start of the
experiment or any other arbitrary time reference (preferingly the epoch)}
}}

