% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_pipeline.R
\name{plot.interaction_table}
\alias{plot.interaction_table}
\title{plot.interaction_table}
\usage{
\method{plot}{interaction_table}(table, OTU_stat, type = "num_int",
  cutoff_type = "q", abundance_type = "mean", ...)
}
\arguments{
\item{table}{An \code{interaction_table}}

\item{OTU_stat}{OTU statistics obtained from the function \link{OTU_stas}}

\item{type}{One of the following: \itemize{
\item \code{'num_int'} Plots the number of interactions for each OTU against
its abundance
\item \code{'ab_prod'} Plots the product of the abundances of each significant pairs of
OTUs against the q- or p-value of the interaction
}}

\item{cutoff_type}{One of the following:
\itemize{
\item \code{'q'} Plots according to the q-values
\item \code{'p'} Plots according to the p-values
}
This parameter is ignored if \code{type='num_int'}}

\item{abundance_type}{The type of abundance to use in the plots. One of \code{c('mean','median','max')}.}

\item{...}{Additional parameters to the \link{plot} function}
}
\description{
Makes a diagnostic plot of an \code{interaction_table}
}
