% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_system.R
\name{predict.LV}
\alias{predict.LV}
\title{Predict trajectory of a Lotka-Volterra system from estimated
coefficients}
\usage{
\method{predict}{LV}(fit, start, times)
}
\arguments{
\item{fit}{An object of class \code{LV_fit} returned from a suitable
fitting function such as \code{\link{ridge_fit}}, containing the Lotka-Volterra coefficients}

\item{start}{The vector of the microbial community at the start
of the simulation. The OTUs must be in the same order as in the
\code{fit} argument.}

\item{times}{The time points to include in the simulation}
}
\value{
Matrix of class \code{deSolve}. For more information, see the description in \link{ode}.
}
\description{
This function utilizes the \code{\link{deSolve}} package in order to
simulate a Lotka-Volterra system with fitted coefficients. Hence, it can be used
for predictions.
}
\seealso{
\link{ridge_fit}, \link{deSolve-package}
}
