% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_pipeline.R
\name{ratio_shared_interactions}
\alias{ratio_shared_interactions}
\title{Finds the Jaccard index of number of shared interactions between the similarity measures}
\usage{
ratio_shared_interactions(similarity_measures_significance)
}
\description{
Finds the Jaccard index of number of shared interactions between the similarity measures
}
