% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{remove_metadata}
\alias{remove_metadata}
\title{remove_metadata}
\usage{
remove_metadata(OTU_table, metadataCols = c("OTU Id", "taxonomy"))
}
\arguments{
\item{OTU_table}{The raw OTU table}

\item{metadataCols}{The names (character vector) or position (integer) of the
metadata columns to remove from the table}
}
\description{
Removes metadata columns from dataset
}
