% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_pipeline.R
\name{runAnalysis}
\alias{runAnalysis}
\title{runAnalysis}
\usage{
runAnalysis(OTU_table, abundance_cutoff = 1e-04, q_crit = 0.05,
  parallel = TRUE, returnVariables = NULL, subset = NULL)
}
\description{
Runs an automized processing of the OTU table, passes the jobs to \code{ccrepe} and saves the results
}
