% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{scale_by_column}
\alias{scale_by_column}
\title{Scale a phyloseq object}
\usage{
scale_by_column(object, column)
}
\arguments{
\item{object}{A \code{phyloseq} object, where the abundances in its OTU table are to be scaled}

\item{column}{Either a character vector of length 1 with the name of the column in the sample data to scale by or a
numeric vector which length is equal to the number of samples.}
}
\value{
A \code{phyloseq} object where the abundances in the OTU table are scaled
}
\description{
This function scales the abundances in each sample by a constant, making it suitable to transform from
relative data to absolute data.
}
\examples{
library(micInt)
library(phyloseq)
data("soilrep")
s <- 1:nsamples(soilrep)
scale_by_column(soilrep, s)
}
