% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_measure_class.R
\docType{class}
\name{sim.measure.attributes-class}
\alias{sim.measure.attributes-class}
\alias{sim.measure.attributes}
\alias{sim.measure.attributes,sim.measure-method}
\alias{sim.measure.attributes,character-method}
\title{Attributes of similarity measures}
\usage{
sim.measure.attributes(x, ...)

\S4method{sim.measure.attributes}{sim.measure}(x)

\S4method{sim.measure.attributes}{character}(x, type_measure, signed)
}
\arguments{
\item{x}{An \link{sim.measure} contaning the information to be displayed or a character string contaning the same of
the similarity measure}

\item{...}{further arguments passed to or from other methods.}

\item{type_measure}{Character, the type of similarity measure which should be repesented}

\item{signed}{Logical, is the similarity measure signed?}
}
\description{
The useful metadata fields for a \code{sim.measure}
}
\section{Slots}{

\describe{
\item{\code{string}}{Character, the similarity score's human readable name}

\item{\code{type_measure}}{Character, what type to similarity measure is this (equvialent to \code{type} field
in \link{sim.measure})}

\item{\code{signed}}{Logical, does this similarity measure return signed results?}
}}

