% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_measure_class.R
\docType{class}
\name{sim.measure.attributes}
\alias{sim.measure.attributes}
\alias{sim.measure.attributes,sim.measure-method}
\alias{sim.measure.attributes,character-method}
\title{sim.measure.attributes}
\usage{
\S4method{sim.measure.attributes}{sim.measure}(x)

\S4method{sim.measure.attributes}{character}(x, type_measure, signed)
}
\description{
The useful metadata fields for a \code{sim.measure}
}
\section{Slots}{

\describe{
\item{\code{string}}{Character, the similarity score's human readable name}

\item{\code{type_measure}}{Character, what type to similarity measure is this (equvialent to \code{type} field
in \link{sim.measure})}

\item{\code{signed}}{Logical, does this similarity measure return signed results?}
}}

