% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_measures.R
\docType{class}
\name{sim.measure}
\alias{sim.measure}
\alias{sim.measure}
\title{sim.measure}
\usage{
sim.measure(FUN, string, categorical = FALSE, mean_scaleable = TRUE)
}
\description{
A similarity measure structure with the function ifself, and metadata fields
}
\section{Slots}{

\describe{
\item{\code{FUN}}{The similarity measure function to use, of the form \code{f(x,y)}.
The function must satisfy the following: \itemize{
\item If \code{x} and \code{y} are both given, they are treated as vectors and the
similarity score between them is returned
\item If only \code{x} is given, it is treated as a matrix (or data frame) where the
features are in columns and sample in the rows. The function must then return
the matrix of pairwise similarities
\item The function must return a similarity score in the interval \eqn{[0,1]}, where 1 is
perfect similarity and 0 is perfect dissimilarity. Alternativly, signed similarity scores
in the interval \eqn{[-1,1]} can also be used.
}}

\item{\code{string}}{The similarity score's human readable name.}

\item{\code{categorical}}{Logical value indicating whether the function only
considers presense-absense of an OTU. If \code{TRUE}, the function will not be
noisified as there is no point in adding noise to the data.}

\item{\code{mean_scaleable}}{Logical value telling whether it makes sense to scale the input by its
mean before applying this function}
}}

\seealso{
noisify
}
