% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_measures.R
\name{similarity_measures}
\alias{similarity_measures}
\title{similarity_measures}
\usage{
similarity_measures(subset = NULL)
}
\description{
Similarity measures passed as functions to ccrepe and
output_ccrepe_data
}
