% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{summary.interaction_table}
\alias{summary.interaction_table}
\title{Create summary of \code{interaction_table}}
\usage{
\method{summary}{interaction_table}(table)
}
\arguments{
\item{table}{An interaction table returned from \link{create_interaction_table}}
}
