% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{test_LV_fit}
\alias{test_LV_fit}
\title{Measure fit of Lotka-Volterra coefficients}
\usage{
test_LV_fit(test_equations, solution_matrix)
}
\arguments{
\item{test_equations}{The list of list of equations to test, should be independent of the \code{solution_matrix}.
For more information, see the description of \code{test_equation} in \link{ridge_fit}}

\item{solution_matrix}{The proposed solution of which the fitness is found, such as the ones returned from
\link{ridge_fit}}
}
\value{
A list of two:
 \itemize{
\item \code{'RMSE'} The root mean square error of the right sides
\item \code{'MAE'} The mean absolute error of the right sides
}
}
\description{
When using cross-validation to find the Lotka-Volterra
coefficients, this function is used to determine the error of the cross-validation. It does
so by applying the
}
\details{
The errors are combined and calculated among all differences between the actual right
sides of the equations and the predicted ones
}
