% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{write.interaction_table}
\alias{write.interaction_table}
\title{write.interaction_table}
\usage{
write.interaction_table(significant_interactions, filename,
  csv_option = "1")
}
\description{
Takes an \code{interaction_table} and writes it to a file. For information about parameters,
see \link{output_ccrepe_data}
}
