% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{write.interactions_table}
\alias{write.interactions_table}
\title{write.interactions_table}
\usage{
write.interactions_table(significant_interactions, filename, csv_option = "2")
}
\description{
Takes an \code{interaction_table} and writes it to a file. For information about parameters,
see \link{output_ccrepe_data}
}
