% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputargs.R
\name{add_outputargs}
\alias{add_outputargs}
\title{Add outputargs}
\usage{
add_outputargs(
  ccrepe_res,
  taxonomy = NULL,
  output.file = NULL,
  return.value = NULL,
  threshold.type = NULL,
  threshold.value = NULL,
  csv_option = NULL,
  removeDuplicates = NULL
)
}
\arguments{
\item{ccrepe_res}{A list of results from \code{\link{ccrepe}}}

\item{taxonomy}{Named character, with the names being those of the OTUs and the values
their taxonomy collapased into a single string (for each OTU)}

\item{output.file}{Should the function write the date to a csv-file?}

\item{return.value}{Should the function return the interaction table as a return value?}

\item{threshold.type}{The type of threshold to be used \code{'q'} denotes q-value,
while \code{'p'} denotes p-value}

\item{threshold.value}{The critical significance value for including an interaction}

\item{csv_option}{If assigned the value \code{'2'}, the function will use the \code{write.csv2} function which print
the numbers if comma as decimal delimer and semicolon as the delimer between numbers.
Else, the decimal delimer is point and comma the delimer between numbers.}

\item{removeDuplicates}{If \code{TRUE}, the function ensures that no OTU pair is listed twice in the results.
Else, two interacting OTUs may show up twice in reverse order
(a pair where OTU A is listed as \code{OTU_1} and OTU B is listed was \code{OTU_2} and vice versa for the other pair)}
}
\description{
Add  output arguments being used by \code{\link{output_ccrepe_data}} in addition to
the ones returned from \code{\link{ccrepe_analysis}}.
}
