% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{as.edge_list.interaction_table}
\alias{as.edge_list.interaction_table}
\alias{as.edgelist.interaction_table}
\title{as.edge_list.interaction_table}
\usage{
\method{as.edgelist}{interaction_table}(table)
}
\description{
Converts an interaction table to a two column character matrix
where each row represent an edge between the OTUs
}
