% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccrepe_analysis.R
\name{ccrepe_analysis}
\alias{ccrepe_analysis}
\title{Conduct ccrepe analysis}
\usage{
ccrepe_analysis(
  ccrepe_job,
  commonargs,
  parallel = FALSE,
  ncpus = getOption("micInt.ncpus", 1L),
  cl = NULL,
  verbose = TRUE
)
}
\arguments{
\item{ccrepe_job}{A list of jobs to be passed to \code{ccrepe}. The lists themselves are named lists with the arguments
being passed to \code{\link{ccrepe}}}

\item{commonargs}{\code{ccrepe} arguments common for all jobs}

\item{parallel}{Should the analysis be run in parallel?}

\item{ncpus}{If \code{parallel = TRUE}, how many cores should be used? Defaults to one.}

\item{cl}{Custom cluster to use if \code{parallel = TRUE}.}

\item{verbose}{Should the function display how much time it spent?}
}
\value{
A list of the results of the various jobs.
Each element of this list containings the ccrepe results
}
\description{
A wrapper around the \code{ccrepe} function, provides parallel analysis
}
\examples{
library(micInt)
data("seawater")
sim.scores <- similarity_measures(subset= c("spearman","pearson"))
sim_funs <- lapply(sim.scores,sim_measure_function)
refined_table <- refine_data(seawater)
ccrepe_commonargs <- list(x = refined_table, iterations = 100,
memory.optimize = TRUE
, min.subj = 5)
ccrepe_job <- list(spearman=list(sim.score = sim_funs[["spearman"]]),
pearson = list(sim.score = sim_funs[["pearson"]]))
ccrepe_analysis(ccrepe_job,ccrepe_commonargs, parallel = TRUE, ncpus = 2)
}
\seealso{
\link{ccrepe}
}
