% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{create_interaction_table}
\alias{create_interaction_table}
\title{create_interaction_table}
\usage{
create_interaction_table(data, OTU_table = NULL, threshold.type = "q",
  threshold.value = 0.05, removeDuplicates = TRUE,
  score_attributes = NULL)
}
\value{
An \code{interactions_table}
}
\description{
Performs the actual process of making the table.
For documentation of parameters, see \link{output_ccrepe_data}
}
