% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputargs.R
\name{create_prefix}
\alias{create_prefix}
\title{create_prefix}
\usage{
create_prefix(
  q_crit = NULL,
  p_crit = NULL,
  cutoff = NULL,
  magfac = NULL,
  digits = 2
)
}
\arguments{
\item{q_crit}{critical q-value. Ignored if not set}

\item{p_crit}{critical p-value, ignored if \code{q-crit} is set.}

\item{cutoff}{The abundance cutoff. Ignored if not set}

\item{magfac}{The magnitude factor. Ignored if not set}

\item{digits}{The number of significant digits in the resulting string}
}
\description{
Standard function to create systematic prefixes from
parameters. The output will be of the form
\code{q_crit=(critical q-value)_cutoff=(the mean abundance cutoff)_magfac=(the magnitude factor)}
in scientific notation.
}
