% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{cut_abundances}
\alias{cut_abundances}
\title{cut_abundances}
\usage{
cut_abundances(refined_table, abundance_cutoff = 0, type = "mean",
  renormalize = TRUE)
}
\description{
Removes metadata columns from dataset
}
