% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{cut_abundances}
\alias{cut_abundances}
\title{cut_abundances}
\usage{
cut_abundances(refined_table, abundance_cutoff = 0, type = "mean",
  renormalize = TRUE)
}
\arguments{
\item{abundance_cutoff}{Numeric, the threshold cutoff value}

\item{type}{The type of measure to base the cutoff on. Can be any
of \code{'mean'}, \code{median}, \code{max} which cuts away
OTUs based on mean, median and maximum abundance, repectivly}

\item{renormalize}{Logical, should the abundances be renormalized after the procedyre?}
}
\description{
Removes metadata columns from dataset
}
