% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_measures.R
\name{noisify}
\alias{noisify}
\title{Noisify}
\usage{
noisify(
  sim.scores = mean_scale(),
  magnitude = 1e-05,
  noise = c("none", "uniform", "normal")
)
}
\arguments{
\item{sim.scores}{The list of objects of class \link{sim.measure} to nosify}

\item{magnitude}{The magnitude of the noise to add. This is: The standard deviation for normal noise and the
radius for the interval for uniform noise}

\item{noise}{The type of noise to add.
\code{'none'} just returns the similarity measure without doing anything.
\code{'uniform'} adds uniformly distributed noise
\code{'normal'} adds normaly distributed  noise
Any combination of the options can be passes as a character vector.
In this case, sim.score functions are returned for the selected types of noise.}
}
\value{
A list of \link{sim.measure} objects corresponding to the input where noise have been
added
}
\description{
Creates similarity measure functions which adds random noise
at certain magnitude to the data before calculating
the similarity score
}
