% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTU_stats.R
\name{OTU_stats}
\alias{OTU_stats}
\title{OTU statistics}
\usage{
OTU_stats(refined_table, taxonomy = NULL)
}
\arguments{
\item{refined_table}{The refined OTU table returned from \link{refine_data}}

\item{taxonomy}{Named character, with the names being those of the OTUs and the values
their taxonomy collapased into a single string (for each OTU)}
}
\value{
A \code{data.frame} with one row for each OTU with the following
elements:

\itemize{
\item \code{ID} The OTU name
\item \code{meanAbundance}
\item \code{medianAbundance}
\item \code{maxAbundance}
\item \code{numberNonZero} The number of samples where the OTU has a
non-zero abundance
\item \code{taxonomy} collapsed into a single string
}
}
\description{
This function returns a summary of OTU statistics for the OTUs of
an OTU table
}
