% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{output_ccrepe_data}
\alias{output_ccrepe_data}
\title{output_ccrepe_data}
\usage{
output_ccrepe_data(
  data,
  taxonomy = NULL,
  threshold.type = "q",
  threshold.value = 0.05,
  output.file = FALSE,
  filename = NULL,
  return.value = TRUE,
  csv_option = "2",
  removeDuplicates = TRUE,
  score_attributes = NULL
)
}
\arguments{
\item{data}{The results from ccrepe}

\item{taxonomy}{Named character, with the names being those of the OTUs and the values
their taxonomy collapased into a single string (for each OTU)}

\item{threshold.type}{The type of threshold to be used \code{'q'} denotes q-value,
while \code{'p'} denotes p-value}

\item{threshold.value}{The critical significance value for including an interaction}

\item{output.file}{Should the function write the date to a csv-file?}

\item{filename}{What is the name of the output file (including path). Ignored if \code{output.file} is \code{TRUE}}

\item{return.value}{Should the function return the interaction table as a return value?}

\item{csv_option}{If assigned the value \code{'2'}, the function will use the \code{write.csv2} function which print
the numbers if comma as decimal delimer and semicolon as the delimer between numbers.
Else, the decimal delimer is point and comma the delimer between numbers.}

\item{removeDuplicates}{If \code{TRUE}, the function ensures that no OTU pair is listed twice in the results.
Else, two interacting OTUs may show up twice in reverse order
(a pair where OTU A is listed as \code{OTU_1} and OTU B is listed was \code{OTU_2} and vice versa for the other pair)}

\item{score_attributes}{An object of class \linkS4class{sim.measure.attributes} belonging to the similarity measure being used}
}
\description{
Takes input from ccrepe and transforms it into a convenient table
}
\examples{
library(micInt)
sim.scores <- similarity_measures(subset= c("spearman","pearson"))
res <- runAnalysis(OTU_table = seawater,sim.scores = sim.scores,parallel = FALSE)
output_ccrepe_data(res$ccrepe_res)


}
