% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot trajectory for time series}
\usage{
plot_trajectory(
  time_series_list,
  distance = "bray",
  subset = names(time_series_list),
  label = FALSE,
  label_size = 3,
  color = NULL,
  linetype = NULL
)
}
\arguments{
\item{time_series_list}{A list of \link{OTU_time_series} objects}

\item{distance}{The name of any distance metric in \link{vegdist}}

\item{subset}{Character, the subsets of time series to plot, based on the
names of the list.}

\item{label}{Logical, should each point be labelled with the time point?}

\item{label_size}{Positive numeric, the font size of the label. Ignored if \code{text=FALSE}.}

\item{color}{Character, the color mapping to use in the plot. Defaults to the time series}

\item{linetype}{Character, the linetype mapping to be used. Defaults to solid}
}
\value{
A \link{ggplot} object showing the trajectory in the two first principal components
}
\description{
When looking at PCA-plots for different time series, we often want to study their evolution
in time. This convenience functions does so by adding arrows between the samples in their
chronological order in the PCA plot
}
\details{
The function utilzes the entire table to make the principal components,
even if some data points are excluded. Note the columns \code{'time_series'} and \code{'time_points'}
which are available in the object being returned.
}
