% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot trajectory for time series}
\usage{
plot_trajectory()
}
\value{
A \link{ggplot} object showing the trajectory in the two first principal components
}
\description{
When looking at PCA-plots for different time series, we often want to study their evolution
in time. This convenience functions does so by adding arrows between the samples in their
chronological order in the PCA plot
}
\details{
This function utilizes
}
