% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_system.R
\name{predict.LV}
\alias{predict.LV}
\title{Predict trajectory of a Lotka-Volterra system from estimated
coefficients}
\usage{
\method{predict}{LV}(fit, start, times, diagnostics = TRUE, ...)
}
\arguments{
\item{fit}{An object of class \code{LV_fit} returned from a suitable
fitting function such as \code{\link{ridge_fit}}, containing the Lotka-Volterra coefficients}

\item{start}{The vector of the microbial community at the start
of the simulation. The OTUs must be in the same order as in the
\code{fit} argument.}

\item{times}{The time points to include in the simulation}

\item{...}{Additional parameters to the function \code{\link{ode}} used to solve the system}
}
\value{
An object of type \code{\link{OTU_time_series}}, where the
table is a matrix of type \code{\link{deSolve}}
}
\description{
This function utilizes the \code{\link{deSolve}} package in order to
simulate a Lotka-Volterra system with fitted coefficients. Hence, it can be used
for predictions.
}
\details{
By default, the \code{lsoda} is the method used
}
\seealso{
\link{ridge_fit}, \link{deSolve-package}
}
