% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{refine_data}
\alias{refine_data}
\title{Refine raw OTU table}
\usage{
refine_data(
  OTU_table,
  abundance_cutoff = 0,
  cutoff_type = "mean",
  renormalize = TRUE,
  metadataCols = c("OTU Id", "taxonomy")
)
}
\arguments{
\item{OTU_table}{The raw OTU table, either as a \code{data.frame}, a \code{matrix} or a \code{phyloseq} object}

\item{abundance_cutoff}{Numeric, the threshold cutoff value. If it is \code{NULL}, the filtering process is skipped.}

\item{cutoff_type}{The type of measure to base the cutoff on. Can be any
of \code{'mean'}, \code{median}, \code{max} which cuts away
OTUs based on mean, median and maximum abundance, repectivly}

\item{renormalize}{Logical, should the abundances be renormalized after the procedyre?}

\item{metadataCols}{The names (character vector) or position (integer) of the
metadata columns to remove from the table}

\item{refined_table}{A \code{data.frame} where OTUs are in columns and samples in rows with metadata removed}
}
\value{
A data frame (always a base data frame even though a tibble is supplied)
with the metadata columns removed, and the OTUs
below the cutoff are filtered away. Additionally, in the refined table
the OTUs constitute the rows, while the rows are the samples (transposed
compared to the original OTU table).
}
\description{
Removes metadata from OTU table and cuts off the least abundant
species, defined by the cutoff parameter
}
\details{
\subsection{Critera for OTU tables}{
In order for an OTU-table to be valid, the following criteria must hold:

\itemize{
\item
The data points (sample) are in columns, the abundances for each
OTU is in rows.
\item
The rows may only hold OTU abundances
\item
There may be as many metadata colums as preferable. However, the all
need to be declared in the \code{metadataCols} argument and the column
\code{taxonomy} has be there in order for the output file to contain the
taxonomy.
\item The row names of the table are the OTU names and the column names are the
sample names
}
Of course, this does not apply to the case when a \code{phyloseq} object is provided as everything
 will be handled automatically
}
\subsection{Tibbles}{
Tibbles are troublesome in this context as the do not support rownames. In order to obtain the OTU IDs, the function
looks through the following in order and proceed with the next until successfull:
\enumerate{
\item If the original OTU table has rownames, they are used as the OTU IDs
\item If one (the first) of \code{c('OTU Id','#OTU ID','OTU ID','OTU_ID')} reside in the colnames of the OTU table,
the corresponding column is used as the OTU IDs
\item If \code{metadataCols} is not empty, the first of the specified columns regarded as the OTU ID (gives warning)
\item The numeric row indicies are treated as the OTU IDs (gives warning)
}
}
}
\examples{
library(micInt)
data("seawater")
refine_data(seawater,cutoff_type = "max")


}
