% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{refine_data}
\alias{refine_data}
\title{refine_data}
\usage{
refine_data(OTU_table, abundance_cutoff = 0, cutoff_type = "mean",
  renormalize = TRUE)
}
\description{
Removes metadata from OTU table and cuts off the least abundant
species, defined by the cutoff parameter
}
