% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_pipeline.R
\name{runAnalysis}
\alias{runAnalysis}
\title{runAnalysis}
\usage{
runAnalysis(OTU_table, abundance_cutoff = 1e-04, q_crit = 0.05,
  parallel = TRUE, returnVariables = NULL, subset = NULL,
  sim.scores = NULL, file = FALSE, magnitude_factor = 10,
  prefix = NULL, metadataCols = c("OTU Id", "taxonomy"),
  postfix = "")
}
\arguments{
\item{OTU_table}{The raw OTU table (if a \code{data.frame} is supplied) to be treated
\emph{or} a \code{phyloseq} object containing the data (the latter is recommended)}

\item{abundance_cutoff}{The mean abundance cutoff for the OTUs. If it is \code{NULL}, the there will be not filtering.}

\item{parallel}{Should the analysis be run in parallel?}

\item{returnVariables}{Which variables should the function return (character vector)?
Available options are: \itemize{
\item \code{similarity_measures_significance}:
The \code{interactions_table} of significant interactions
\item \code{refined_table}: The processed  OTU table
\item \code{min_dataset}: The smallest non-zero entity in the refined table
\item \code{taxonomy}: A named numberic containing the taxonomy of each OTU (collapsed into a single string)
}
In addition, all paramerters for this function are available.
If \code{NULL}, all variables in the function will be returned.
Note: If \code{OTU_table} is a \code{phyloseq} object, the returned variables is a data frame corresponding to the
\code{phyloseq} object. This is due to the fact that it is intervally converted into a data frame.}

\item{subset}{Character, the subset of similarity measures to use, denoted by
the its name in the list (not necessarly its string) returned from \link{similarity_measures} or similarity
measure modiftying function such as \link{noisify}
If \code{NULL}, all available measures will be used}

\item{sim.scores}{The similarity measures of class \link{sim.measure}
to use. If it is \code{NULL}, all measures available in the package
will be used (recommanded for most purposes).}

\item{file}{Should the tables of significant interactions be printed to a file?}

\item{magnitude_factor}{When making noisified functions, the magnitude of the noise
will be this number multiplied with \code{min_dataset}}

\item{prefix}{The prefix of the file names being written. Ignored if \code{file=FALSE}.}

\item{metadataCols}{The names (character vector) or position (integer) of the
metadata columns to remove from the table before analyzing it. Ignored if a \code{phyloseq} object is supplied}
}
\value{
A list of the variables requested from the parameter \code{returnVariables}.
}
\description{
Runs an automized processing of the OTU table or phyloseq object, passes the jobs to \code{ccrepe} and saves the results
}
\details{
If the function is told to output a file and no prefix is given, the csv-files will all share a common prefix of the form:
\code{q_crit=(critical q-value)_cutoff=(the mean abundance cutoff)_magfac=(the magnitude factor)},
where all numbers are in scientific notation. Then the sim.score name follows, then the postfix and finally the csv
extention.
The postfix is by default empty.

In order for an OTU-table to be valid when the argument \code{OTU_table} is a \code{data.frame},
the following criteria must hold:

\itemize{
\item
The data points (sample) are in columns, the abundances for each
OTU is in rows.
\item
The rows may only hold OTU abundances
\item
There may be as many metadata colums as preferable. However, the all
need to be declared in the \code{metadataCols} argument and the column
\code{taxonomy} has be there in order for the output file to contain the
taxonomy.
\item The row names of the table are the OTU names and the column names are the
sample names
}
For \code{phyloseq} object, you do not need to care about this, it is automatically handeled
}
