% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_pipeline.R
\name{runAnalysis}
\alias{runAnalysis}
\title{runAnalysis}
\usage{
runAnalysis(OTU_table, abundance_cutoff = 1e-04, q_crit = 0.05,
  parallel = TRUE, returnVariables = NULL, subset = NULL,
  sim.scores = NULL, file = FALSE, magnitude_factor = 10, prefix = NULL,
  postfix = "")
}
\arguments{
\item{OTU_table}{The raw OTU table to be treated}

\item{abundance_cutoff}{The mean abundance cutoff for the OTUs}

\item{parallel}{Should the analysis be run in parallel?}

\item{returnVariables}{Which variables should the function return (character vector)?
Available options are: \itemize{
\item \code{similarity_measures_significance}:
The \code{interactions_table} of significant interactions
\item \code{refined_table}: The processed  OTU table
\item \code{min_dataset}: The smallest non-zero entity in the refined table
}
In addition, all paramerters for this function are available
If \code{NULL}, all variables in the function will be returned}

\item{subset}{The subset of similarity measures to use}

\item{sim.scores}{The similarity measures of class \link{sim.measure}
to use. If it is \code{NULL}, all measures available in the package
will be used (recommanded for most purposes).}

\item{file}{Should the tables of significant interactions be printed to a file?}

\item{magnitude_factor}{When making noisified functions, the magnitude of the noise
will be this number multiplied with \code{min_dataset}}

\item{prefix}{The prefix of the file names being written. Ignored if \code{file=FALSE}.}
}
\description{
Runs an automized processing of the OTU table, passes the jobs to \code{ccrepe} and saves the results
}
\details{
If the function is told to output a file and no prefix is given, the csv-files will all share a common prefix of the form:
\code{q_crit=(critical q-value)_cutoff=(the mean abundance cutoff)_magfac=(the magnitude factor)},
where all numbers are in scientific notation. Then the sim.score name follows and then the postfix
The postfix is by default empty.
}
