% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_pipeline.R
\name{runAnalysis}
\alias{runAnalysis}
\title{runAnalysis}
\usage{
runAnalysis(
  OTU_table,
  abundance_cutoff = 1e-04,
  q_crit = 0.05,
  parallel = FALSE,
  ncpus = getOption("micInt.ncpus", 1L),
  cl = NULL,
  returnVariables = NULL,
  subset = NULL,
  sim.scores = NULL,
  file = FALSE,
  magnitude_factor = 10,
  prefix = NULL,
  metadataCols = c("OTU Id", "taxonomy"),
  postfix = "",
  renormalize = TRUE,
  iterations = 1000,
  ccrepe_args = list()
)
}
\arguments{
\item{OTU_table}{The raw OTU table (if a \code{data.frame} or a \code{phyloseq} \code{otu_table} is supplied) to be treated
\emph{or} an experiment level \code{phyloseq} object containing the data (the latter is recommended).
Note that in the case of a \code{phyloseq} \code{otu_table}, no taxonomy can be handled.}

\item{abundance_cutoff}{The mean abundance cutoff for the OTUs. If it is \code{NULL}, the there will be not filtering.}

\item{q_crit}{Numeric, the q-value cutoff when construction interaction tables}

\item{parallel}{Should the analysis be run in parallel?}

\item{ncpus}{If \code{parallel = TRUE}, how many cores should be used? Defaults to one.}

\item{cl}{Custom cluster to use if \code{parallel = TRUE}.}

\item{returnVariables}{Which variables should the function return (character vector)?
Available options are: \itemize{
\item \code{similarity_measures_significance}:
The \code{interaction_table} of significant interactions
\item \code{refined_table}: The processed  OTU table
\item \code{min_dataset}: The smallest non-zero entity in the refined table
\item \code{taxonomy}: A named numberic containing the taxonomy of each OTU (collapsed into a single string)
\item \code{outputargs}: A list (with a element for each similarity measure) comtaining the arguments to be passed
to \link{output_ccrepe_data} for each similarity measure
\item \code{common_outputargs}: Like \code{outputargs}, but these arguments stay the same for all similarity
measure in order to avoid duplicates.
}

In addition, all paramerters for this function are available. Other internal variables
found upon inspection of the source code may also be returned,
but they are for advanced users only.
If \code{NULL}, the listed parameters in this section in addtion to an echo of the parameters are retuned.
Note: If \code{OTU_table} is a \code{phyloseq} object, the returned variables is a data frame corresponding to the
\code{phyloseq} object. This is due to the fact that it is intervally converted into a data frame.}

\item{subset}{Character, the subset of similarity measures to use, denoted by
the its name in the list (not necessarly its string) returned from \link{similarity_measures} or similarity
measure modiftying function such as \link{noisify}
If \code{NULL}, all available measures will be used}

\item{sim.scores}{The similarity measures of class \link{sim.measure}
to use. If it is \code{NULL}, all measures available in the package
will be used (recommanded for most purposes).}

\item{file}{Should the tables of significant interactions be written to a file? If so, they are printed to \code{csv}-files containing the name of the similarity measure}

\item{magnitude_factor}{When making noisified functions, the magnitude of the noise
will be this number multiplied with \code{min_dataset}}

\item{prefix}{The prefix of the file names being written. Ignored if \code{file=FALSE}.}

\item{metadataCols}{The names (character vector) or position (integer) of the
metadata columns to remove from the table before analyzing it. Ignored if a \code{phyloseq} object is supplied}

\item{postfix}{The postfix of the file names being written. Ignored if \code{file=FALSE}.}

\item{renormalize}{Should the data be renormalized during filtering process and permutation?
Should be \code{TRUE} when used on relative abundances, but must be \code{FALSE} if absolute abundances are used.}

\item{iterations}{Integer of length one, the number of iterations to run}

\item{ccrepe_args}{A named list of custom arguments to \code{\link{ccrepe}} if it is necessary to fine-tune the workings. This argument list
will override the effects of the other arguments.}
}
\value{
A list of the variables requested from the parameter \code{returnVariables}.
}
\description{
Runs an automized processing of the OTU table or phyloseq object, passes the jobs to \code{ccrepe} and saves the results
}
\details{
If the function is told to output a file and no prefix is given, the csv-files will all share a common prefix of the form:
\code{q_crit=(critical q-value)_cutoff=(the mean abundance cutoff)_magfac=(the magnitude factor)},
where all numbers are in scientific notation. Then the sim.score name follows, then the postfix and finally the csv
extention.
The postfix is by default empty.

In order for an OTU-table to be valid when the argument \code{OTU_table} is a \code{data.frame},
the following criteria must hold:

\itemize{
\item
The data points (sample) are in columns, the abundances for each
OTU is in rows.
\item
The rows may only hold OTU abundances
\item
There may be as many metadata colums as you like. However, they all
need to be declared in the \code{metadataCols} argument and the column
\code{taxonomy} has be there in order for the output file to contain the
taxonomy.
\item The row names of the table are the OTU names and the column names are the
sample names
}
For \code{phyloseq} objects (both experiment level and \code{otu_table}), you do not need to care about this, it is automatically handeled
}
\examples{
library(micInt)
data(seawater)
sim.scores <- similarity_measures(subset= c("spearman","pearson"))
runAnalysis(OTU_table = seawater, sim.scores = sim.scores, parallel = TRUE, ncpus = 2,
iterations = 100)

}
\seealso{
\link{output_ccrepe_data}
}
