% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_pipeline.R
\name{significanceDiganostics}
\alias{significanceDiganostics}
\title{significanceDiganostics}
\usage{
significanceDiganostics(similarity_measures_significance, OTU_table,
  refined_table, metadataCols = NULL, threshold.type = "q")
}
\arguments{
\item{similarity_measures_significance}{A list over interaction tables, see \link{output_ccrepe_data}.}

\item{OTU_table}{The raw OTU table to be treated}

\item{refined_table}{The refined table returned from \link{}}

\item{metadataCols}{The names (character vector) or position (integer) of the
metadata columns to remove from the table before analyzing it}

\item{threshold.type}{The type of threshold to be used \code{'q'} denotes q-value,
while \code{'p'} denotes p-value}
}
\description{
Makes interaction density plots and abundance
product plots for significant interations
}
