% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_measures.R
\name{similarity_measures}
\alias{similarity_measures}
\title{Get premade similarity measure objects}
\usage{
similarity_measures(
  subset = c("pearson", "spearman", "kendall", "bray_curtis", "jaccard", "gen_jaccard",
    "mutual_information", "nc.score", "euclidean", "cosine")
)
}
\arguments{
\item{subset}{The subset of the similarity measures to be returned.
By default, all similarity measures are returned}
}
\value{
A list of the \link{sim.measure} objects defined by the function
}
\description{
This function allows retriving the ten ready-made similarity measure
provided by this package. Custom similarity measures can also be made \link{sim.measure-class}.
}
\examples{
library(micInt)
x <- rnorm(10,mean = 1, sd = 1)
y <- 0.5 * x + 0.5 * rnorm(10, mean =0,sd = sqrt(0.5))
sim_measures <- similarity_measures(subset = c("pearson","spearman"))
pearson_cor <- sim_measures[["pearson"]]
pearson_cor




}
\seealso{
sim.measure-class
}
