% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integralSystem.R
\name{stack_equations}
\alias{stack_equations}
\title{Stack equations from different time series}
\usage{
stack_equations(equations)
}
\arguments{
\item{equations}{A list of lists of equations, where the second list layer contains the
name of the respective OTUs}
}
\description{
For each OTU, all equations will be stacked together
}
