% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_refinement.R
\name{write.interaction_table}
\alias{write.interaction_table}
\title{write.interaction_table}
\usage{
write.interaction_table(significant_interactions, filename, csv_option = "2")
}
\arguments{
\item{significant_interactions}{An \code{interaction_table} returned from \code{\link{create_interaction_table}}}

\item{filename}{What is the name of the output file (including path). Ignored if \code{output.file} is \code{TRUE}}

\item{csv_option}{If assigned the value \code{'2'}, the function will use the \code{write.csv2} function which print
the numbers if comma as decimal delimer and semicolon as the delimer between numbers.
Else, the decimal delimer is point and comma the delimer between numbers.}
}
\description{
Takes an \code{interaction_table} and writes it to a file. For information about parameters,
see \link{output_ccrepe_data}
}
